refactor(schema): split tables, enums, and relations into separate folders with centralized exports
