+++ src/database/schema/tables/index.ts
@@ -1,0 +1,13 @@
+export * from './audit';
+export * from './bulkInvite';
+export * from './image';
+export * from './intervention';
+export * from './migration';
+export * from './notification';
+export * from './project';
+export * from './site';
+export * from './species';
+export * from './survey';
+export * from './tree';
+export * from './user';
+export * from './workspace';
